moves invidual prepare steps to bake-content
